Merge ds/custom_editor into master
--------------------
Commit message for microsoft/vscode-python@b699559775c7662acefe016f2bc171bb12e561f0:

Fix merge issue

--------------------
Commit message for microsoft/vscode-python@eceea7720e059d02c9c6f8678b08d55129375d79:

Merge branch 'master' into ds/custom_editor

* master:
  Move redux logging to a different file (microsoft/vscode-python#10605)
  Ensure local host only if connection not available (microsoft/vscode-python#10600)
  Fix a bunch of debugger issues (microsoft/vscode-python#10572)
  Bump acorn from 6.0.0 to 6.4.1 (microsoft/vscode-python#10570)
  Resize plot to fit within pdf page (microsoft/vscode-python#10547)
  Update dev build instructions to mention "python.insidersChannel"
  Install the ZMQ prebuilt binaries and verify ZMQ works on startup (microsoft/vscode-python#10551)

--------------------
Commit message for microsoft/vscode-python@81a8d7ddcf3439113b6c4366df711e2da08c1982:

Disable custom editor API (microsoft/vscode-python#10604)


--------------------
Commit message for microsoft/vscode-python@f7bf3708e3dffc2f233b07cd9bcb41dc30695be8:

Delegate management of notebooks (creation and disposing) to I… (microsoft/vscode-python#10568)

* No need to dispose
* Refactor
* Remove nb
* Oops
* Add missing imports
* Skip tests
* Fix linter
--------------------
Commit message for microsoft/vscode-python@cf64598376292b3798c5962948930379b6cdea99:

Refactor to accommodate upstream VSCode API changes (microsoft/vscode-python#10569)

* Accomodate upstream VSC changes
* Use weekmap
* Ignore
--------------------
Commit message for microsoft/vscode-python@c9f315f4bd585ba8c359c9835b3266d6842a2fbe:

Merge branch 'master' into ds/custom_editor

* master:
  Rename notebookProvider to notebookEditorProvider (microsoft/vscode-python#10564)

--------------------
Commit message for microsoft/vscode-python@d3c4ad953d9a5c42c6f272b3fc2c1f27e5768140:

Merge branch 'master' into ds/custom_editor

* master:
  Fix flakey file system tests (microsoft/vscode-python#10541)
  Tweaks for gather (microsoft/vscode-python#10532)
  Fix microsoft/vscode-python#10437: Update launch.json handling to support "listen" and "connect" (microsoft/vscode-python#10517)
  Add conda support to nightly flake test (microsoft/vscode-python#10523)
  Rename datascience to datascience_modules (microsoft/vscode-python#10525)
  Clean up the extension activation code. (microsoft/vscode-python#10455)
  Allow escape and ctrl+U to clear the interactive window (microsoft/vscode-python#10513)
  Fix builtins so they don't interfere with our execution (microsoft/vscode-python#10511)
  Jupyter autocompletion will only show up on empty lines, (microsoft/vscode-python#10420)
  notify on missing kernel and interpreter with kernel (microsoft/vscode-python#10508)

--------------------
Commit message for microsoft/vscode-python@cc9458ac16f89e3d90da9ab02bb76bdab6920785:

Use a single notebook beetween multiple native editors (microsoft/vscode-python#10514)

* Share notebook with multiple native editors
* Oops
* Linter fixes
* Disable test
* Fix tests
* Fix linter
--------------------
Commit message for microsoft/vscode-python@d8de1593c32a9484ba4806cb391f7a9bb9ae7809:

Add support for synchronising edits between two nb editors of… (microsoft/vscode-python#10189)

* Add support for syncing edits
* Allow syncing edits and disabled in old notebook
* Fix formatting
* Fix linter issues
--------------------
Commit message for microsoft/vscode-python@54b99ff2c02ff76ffc56cd4c65177eccae0652cc:

Fix tests in ds/custom_editor branch (microsoft/vscode-python#10512)

* Fix tests
* Revert "Disable proposed api used for custom editor"
This reverts commit e93f77ab11e68aeeeb81f82b0632cefd24cecef5.
* More fixes
* Don't need that
--------------------
Commit message for microsoft/vscode-python@88dd250eb3b2f0e70da2e7306a953e25f259e071:

Merge branch 'custom_editor_update' into ds/custom_editor

* custom_editor_update:
  Fix linter
  Reimplement custom editor using new API

--------------------
Commit message for microsoft/vscode-python@a0b6942c15201bcc5f00f6fcca6259a4fb414192:

Fix linter

--------------------
Commit message for microsoft/vscode-python@848cafa485727a3a513fafcfcde2de22bdb81fd0:

Merge remote-tracking branch 'origin/ds/custom_editor' into rchiodo/custom_editor_update

--------------------
Commit message for microsoft/vscode-python@a73b9179df8818639bf7606f04a5a672f2ed37c7:

Reimplement custom editor using new API


Lead-authored-by: Don Jayamanne <don.jayamanne@yahoo.com>

+++ extensions/positron-python/src/client/datascience/interactive-ipynb/nativeEditorStorage.ts
@@ -3,7 +3,7 @@ import * as fastDeepEqual from 'fast-deep-equal';
 import { inject, injectable, named } from 'inversify';
 import * as path from 'path';
 import * as uuid from 'uuid/v4';
-import { Event, EventEmitter, Memento, Uri } from 'vscode';
+import { CancellationToken, Event, EventEmitter, Memento, Uri } from 'vscode';
 import { concatMultilineStringInput, splitMultilineString } from '../../../datascience-ui/common';
 import { createCodeCell } from '../../../datascience-ui/common/cellFactory';
 import { traceError } from '../../common/logger';
@@ -20,6 +20,8 @@ import { CellState, ICell, IJupyterExecution, IJupyterKernelSpec, INotebookModel
 // tslint:disable-next-line:no-require-imports no-var-requires
 import detectIndent = require('detect-indent');
 import { sendTelemetryEvent } from '../../telemetry';
+// tslint:disable-next-line:no-require-imports no-var-requires
+const debounce = require('lodash/debounce') as typeof import('lodash/debounce');
 
 const KeyPrefix = 'notebook-storage-';
 const NotebookTransferKey = 'notebook-transfered';
@@ -50,7 +52,12 @@ export class NativeEditorStorage implements INotebookModel, INotebookStorage {
     public get cells(): ICell[] {
         return this._state.cells;
     }
+    public get onDidEdit(): Event<NotebookModelChange> {
+        return this._editEventEmitter.event;
+    }
+
     private _changedEmitter = new EventEmitter<NotebookModelChange>();
+    private _editEventEmitter = new EventEmitter<NotebookModelChange>();
     private _state: INativeEditorStorageState = {
         file: Uri.file(''),
         changeCount: 0,
@@ -59,6 +66,7 @@ export class NativeEditorStorage implements INotebookModel, INotebookStorage {
         notebookJson: {}
     };
     private indentAmount: string = ' ';
+    private debouncedWriteToStorage = debounce(this.writeToStorage.bind(this), 250);
 
     constructor(
         @inject(IJupyterExecution) private jupyterExecution: IJupyterExecution,
@@ -79,11 +87,17 @@ export class NativeEditorStorage implements INotebookModel, INotebookStorage {
         this.handleModelChange(change);
     }
 
-    public save(): Promise<INotebookModel> {
+    public async applyEdits(edits: readonly NotebookModelChange[]): Promise<void> {
+        edits.forEach(e => this.update({ ...e, source: 'redo' }));
+    }
+    public async undoEdits(edits: readonly NotebookModelChange[]): Promise<void> {
+        edits.forEach(e => this.update({ ...e, source: 'undo' }));
+    }
+    public save(): Promise<void> {
         return this.saveAs(this.file);
     }
 
-    public async saveAs(file: Uri): Promise<INotebookModel> {
+    public async saveAs(file: Uri): Promise<void> {
         const contents = await this.getContent();
         await this.fileSystem.writeFile(file.fsPath, contents, 'utf-8');
         if (this.isDirty || file.fsPath !== this.file.fsPath) {
@@ -96,9 +110,11 @@ export class NativeEditorStorage implements INotebookModel, INotebookStorage {
                 oldDirty: this.isDirty
             });
         }
-        return this;
     }
-
+    public async backup(cancellation: CancellationToken): Promise<void> {
+        // Should send to extension context storage path
+        return this.storeContentsInHotExitFile(cancellation);
+    }
     public async getJson(): Promise<Partial<nbformat.INotebookContent>> {
         await this.ensureNotebookJson();
         return this._state.notebookJson;
@@ -113,7 +129,7 @@ export class NativeEditorStorage implements INotebookModel, INotebookStorage {
      * Also keep track of the current time. This way we can check whether changes were
      * made to the file since the last time uncommitted changes were stored.
      */
-    public async storeContentsInHotExitFile(): Promise<void> {
+    public async storeContentsInHotExitFile(cancelToken?: CancellationToken): Promise<void> {
         const contents = await this.getContent();
         const key = this.getStorageKey();
         const filePath = this.getHashedFileName(key);
@@ -123,15 +139,19 @@ export class NativeEditorStorage implements INotebookModel, INotebookStorage {
         const specialContents = contents ? JSON.stringify({ contents, lastModifiedTimeMs: Date.now() }) : undefined;
 
         // Write but debounced (wait at least 250 ms)
-        return this.writeToStorage(filePath, specialContents);
+        return this.debouncedWriteToStorage(filePath, specialContents, cancelToken);
     }
-    private async writeToStorage(filePath: string, contents?: string): Promise<void> {
+    private async writeToStorage(filePath: string, contents?: string, cancelToken?: CancellationToken): Promise<void> {
         try {
-            if (contents) {
-                await this.fileSystem.createDirectory(path.dirname(filePath));
-                return this.fileSystem.writeFile(filePath, contents);
-            } else {
-                return this.fileSystem.deleteFile(filePath);
+            if (!cancelToken?.isCancellationRequested) {
+                if (contents) {
+                    await this.fileSystem.createDirectory(path.dirname(filePath));
+                    if (!cancelToken?.isCancellationRequested) {
+                        return this.fileSystem.writeFile(filePath, contents);
+                    }
+                } else {
+                    return this.fileSystem.deleteFile(filePath);
+                }
             }
         } catch (exc) {
             traceError(`Error writing storage for ${filePath}: `, exc);
@@ -178,6 +198,15 @@ export class NativeEditorStorage implements INotebookModel, INotebookStorage {
         // Forward onto our listeners if necessary
         if (changed || this.isDirty !== oldDirty) {
             this._changedEmitter.fire({ ...change, newDirty: this.isDirty, oldDirty });
+        }
+        // Slightly different for the event we send to VS code. Skip version and file changes. Only send user events.
+        if (
+            (changed || this.isDirty !== oldDirty) &&
+            change.kind !== 'version' &&
+            change.kind !== 'file' &&
+            change.source === 'user'
+        ) {
+            this._editEventEmitter.fire(change);
         }
     }
